Deactivated the numeric limit of Layer info
Deactivated  the numeric limit (by decimal format) of the Layer.class info related to geographic extension. The previus commit doesn't giv e exact values if  the image is in geographic coordinates (ex. EPSG4326).
With this patch, the accuracy of the infomation, is retrieve directly from the values of the Envelope using JTS. 
As explained   in a former request for RasterImageLayer (#90), this correction can be useful to get information from a Image (in this case loaded via Layer.class) in order to manually modifyr tags into a KML file to obtain the correct position.