Extended XBasicStyle to include al LineStringStyle
This is an experimental class and could be consider a proposal (for fuure works on Styling)
The idea is to extend XBasicStyle, used by Color Theming classification for vector layer, also the class that control the line decoration.
Actally the line decoration implemented in OpenJUMP has an effect on the entire layer and cannot be customized according to a theming classification (by an attribute). 
This implementation requires tests and, of coarse, other jobs on the GUI . And a lot of job to avoid to break backward compatibility.