Update AddRasterImageLayerWizard.java
This part of code allows to read and apply a style to a RasterImageLayer on loading.  The style must be stored as SLD file with the same name of the layer. Limited to DTM/DEM (monoband raster files) as OpenJUMP can actually read/write symbology only for  these types of file