Update AddRasterImageLayerWizard.java
This patch allows to read and apply a style to a RasterImageLayer on loading file.  The style must be stored as SLD file with the same name of the layer.  It is currently limited to single banded layer (DTM/DEM) as SLDHandler class can decode styles only for that type of layers