Allowed ability to automatically apply a style to the raster from a SLD file
This part of code allows to read and apply a style to a RasterImageLayer. The style must be stored as SLD file, with the same name of the raster layer. This ability is limited to single banded layers (DEM/DTM) as   currently OpenJUMP can read/write symbologies only for these types of file